+++ Naive/src/StepNaive.h
@@ -6,14 +6,14 @@
 #include "Step.h"
 
 class StepNaive  : public Step {
-    std::vector<float> forces;
-    std::vector<float> velocities;
-    std::vector<float> weights;
+    std::vector<double> forces;
+    std::vector<double> velocities;
+    std::vector<double> weights;
 public:
     StepNaive(std::vector<float>& masses, unsigned N);
     ~StepNaive();
     void initialize();
-    void compute(tf3& positions, float dt);
+    void compute(tf3& positions, double dt);
 };
 
 #endif
